Revert "add new instrumentation and cover more constructors"
This reverts commit 5dc237b54dce05a0d8131f579c2f9672676ef9c5.
