+++ dd-java-agent/instrumentation/java-io/src/main/java/datadog/trace/instrumentation/java/lang/InputStreamReaderCallSite.java
@@ -12,7 +12,6 @@ import javax.annotation.Nonnull;
 @CallSite(spi = IastCallSites.class)
 public class InputStreamReaderCallSite {
 
-  @CallSite.After("void java.io.InputStreamReader.<init>(java.io.InputStream)")
   @CallSite.After(
       "void java.io.InputStreamReader.<init>(java.io.InputStream, java.nio.charset.Charset)")
   public static InputStreamReader afterInit(
