+++ dd-java-agent/instrumentation/java-io/src/test/java/foo/bar/TestInputStreamReaderSuite.java
@@ -9,8 +9,4 @@ public class TestInputStreamReaderSuite {
   public static InputStreamReader init(final InputStream in, Charset charset) {
     return new InputStreamReader(in, charset);
   }
-
-  public static InputStreamReader init(final InputStream in) {
-    return new InputStreamReader(in);
-  }
 }
