+++ dd-java-agent/instrumentation/org-json/build.gradle
@@ -16,7 +16,6 @@ dependencies {
   testImplementation group: 'org.json', name: 'json', version: '20230227'
 
   testRuntimeOnly project(':dd-java-agent:instrumentation:iast-instrumenter')
-  testRuntimeOnly project(':dd-java-agent:instrumentation:java-io') //Needed for Reader
 
   //FIXME: ASM
   latestDepTestImplementation group: 'org.json', name: 'json', version: '+'
