+++ dd-java-agent/instrumentation/org-json/src/main/java/datadog/trace/instrumentation/json/JSONArrayInstrumentation.java
@@ -5,7 +5,6 @@ import static net.bytebuddy.matcher.ElementMatchers.isConstructor;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
 import static net.bytebuddy.matcher.ElementMatchers.isPublic;
 import static net.bytebuddy.matcher.ElementMatchers.returns;
-import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
 import com.google.auto.service.AutoService;
@@ -15,6 +14,8 @@ import datadog.trace.api.iast.InstrumentationBridge;
 import datadog.trace.api.iast.Propagation;
 import datadog.trace.api.iast.propagation.PropagationModule;
 import net.bytebuddy.asm.Advice;
+import org.json.JSONArray;
+import org.json.JSONObject;
 
 @AutoService(InstrumenterModule.class)
 public class JSONArrayInstrumentation extends InstrumenterModule.Iast
@@ -32,11 +33,11 @@ public class JSONArrayInstrumentation extends InstrumenterModule.Iast
   @Override
   public void methodAdvice(MethodTransformer transformer) {
     transformer.applyAdvice(
-        isConstructor().and(takesArguments(0)).and(takesArgument(0, named("org.json.JSONTokener"))),
+        isConstructor().and(takesArguments(String.class)),
         getClass().getName() + "$ConstructorAdvice");
     transformer.applyAdvice(
         isMethod().and(isPublic()).and(returns(Object.class)).and(named("opt")),
-        packageName + ".OptAdvice");
+        getClass().getName() + "$OptAdvice");
   }
 
   public static class ConstructorAdvice {
@@ -46,6 +47,26 @@ public class JSONArrayInstrumentation extends InstrumenterModule.Iast
       final PropagationModule iastModule = InstrumentationBridge.PROPAGATION;
       if (iastModule != null && input != null) {
         iastModule.taintObjectIfTainted(self, input);
+      }
+    }
+  }
+
+  public static class OptAdvice {
+    @Advice.OnMethodExit(suppress = Throwable.class)
+    @Propagation
+    public static void afterMethod(@Advice.This Object self, @Advice.Return final Object result) {
+      boolean isString = result instanceof String;
+      boolean isJson = !isString && (result instanceof JSONObject || result instanceof JSONArray);
+      if (!isString && !isJson) {
+        return;
+      }
+      final PropagationModule iastModule = InstrumentationBridge.PROPAGATION;
+      if (iastModule != null) {
+        if (isString) {
+          iastModule.taintStringIfTainted((String) result, self);
+        } else {
+          iastModule.taintObjectIfTainted(result, self);
+        }
       }
     }
   }
