+++ dd-java-agent/instrumentation/org-json/src/test/groovy/JSONCookieInstrumentationTest.groovy
@@ -23,8 +23,7 @@ class JSONCookieInstrumentationTest extends AgentTestRunner {
 
     then:
     1 * module.taintObjectIfTainted(_ as JSONObject, cookie)
-    1 * module.taintObjectIfTainted(_ as Reader, cookie)
-    1 * module.taintObjectIfTainted(_ as JSONTokener, _ as Reader)
+    1 * module.taintObjectIfTainted(_ as JSONTokener, cookie)
     0 * _
   }
 }
